aya-ebpf: use correct types for BPF helper return values
Complete the i32 errno migration started in 8962fc79. Errno-only
helpers now return Result<T, i32>, non-errno helpers (verdicts, byte
counts) return c_long, and internal functions receiving raw helper
returns use c_long. check_bounds_signed now takes T: Into<i64> to
accept all signed/unsigned integer types without manual casts.

Fixes: #1414

+++ ebpf/aya-ebpf/src/maps/ring_buf.rs
@@ -192,7 +192,7 @@ impl RingBuf {
     ///
     /// [`reserve`]: RingBuf::reserve
     /// [`submit`]: RingBufEntry::submit
-    pub fn output<T: ?Sized>(&self, data: impl Borrow<T>, flags: u64) -> Result<(), i64> {
+    pub fn output<T: ?Sized>(&self, data: impl Borrow<T>, flags: u64) -> Result<(), i32> {
         let data = data.borrow();
         assert_eq!(8 % align_of_val(data), 0);
         let ret = unsafe {
@@ -203,7 +203,7 @@ impl RingBuf {
                 flags,
             )
         };
-        if ret < 0 { Err(ret) } else { Ok(()) }
+        if ret < 0 { Err(ret as i32) } else { Ok(()) }
     }
 
     /// Query various information about the ring buffer.
